+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -21,3 +21,4 @@
 * Implements image streaming.
 * Provides LifecycleOwner implementation for Activities that use the plugin that do not implement it themselves.
 * Implements retrieval of camera information.
+* Implements video capture.